requirements.txt
python-telegram-bot==20.7
pillow
pytesseract
python-dotenv
